move bip327 tests to separate file
